controllers/krate/metadata: Move `GET /summary` endpoint to a dedicated module
This module is focussed on the `GET /crates/:crate_id` endpoint and its children. The `GET /summary` endpoint is a distant relative, but doesn't really belong in this module.
