+++ src/router.rs
@@ -122,7 +122,7 @@ pub fn build_axum_router(state: AppState) -> Router<()> {
             "/api/v1/me/email_notifications",
             put(user::me::update_email_notifications),
         )
-        .route("/api/v1/summary", get(krate::metadata::summary))
+        .route("/api/v1/summary", get(summary::summary))
         .route(
             "/api/v1/confirm/:email_token",
             put(user::me::confirm_user_email),
